internalize observer event test hook
