+++ packages/runtime/src/session/session-plugin-events.test.ts
@@ -5,13 +5,10 @@ import {
   assistantMessage,
   createCallbackModel,
   eventTypes,
-  sentUserText,
   userText,
 } from "../test-fixtures";
 import { userTextToModelMessage } from "./mapping";
-import { AgentSession } from "./session";
 import { collect } from "./session.test-support";
-import { MemorySessionStore } from "./store/memory";
 
 describe("Agent session plugin events", () => {
   it("calls event plugins for queued turn events", async () => {
@@ -70,45 +67,6 @@ describe("Agent session plugin events", () => {
     const events = await collect(await agent.send("hello"));
 
     expect(pluginEventTypes).toEqual(eventTypes(events));
-  });
-
-  it("routes observer events through event plugins", async () => {
-    const pluginEventTypes: string[] = [];
-    const session = new AgentSession(
-      {
-        model: createCallbackModel(() =>
-          Promise.resolve([assistantMessage("DONE")])
-        ),
-      },
-      { key: "observer-events", store: new MemorySessionStore() },
-      [
-        {
-          on: ({ event }) => {
-            pluginEventTypes.push(event.type);
-          },
-        },
-      ]
-    );
-
-    const run = await session.send("hello");
-    const iterator = run.events()[Symbol.asyncIterator]();
-
-    expect((await iterator.next()).value).toEqual(sentUserText("hello"));
-    expect((await iterator.next()).value).toEqual({ type: "turn-start" });
-
-    await session.emitObserverEvent({
-      text: "observer reasoning",
-      type: "assistant-reasoning",
-    });
-
-    const events = [
-      (await iterator.next()).value,
-      (await iterator.next()).value,
-    ];
-    await iterator.return?.();
-
-    expect(eventTypes(events)).toContain("assistant-reasoning");
-    expect(pluginEventTypes).toContain("assistant-reasoning");
   });
 
   it("commits successful output before terminal event plugin failures", async () => {
